17-5-27 backup

+++ src/main/java/project/basic/controller/BookClassController.java
@@ -14,6 +14,7 @@ import project.basic.model.ClassList;
 import project.navigator.route.Forms;
 import project.navigator.route.Lists;
 import project.navigator.service.CacheManager;
+import project.operation.entity.Book;
 import project.system.entity.AdminLog;
 import project.system.model.AdminSession;
 import project.system.pojo.OperationTargets;
@@ -38,14 +39,15 @@ public class BookClassController {
     @Autowired
     private CacheManager cacheManager;
 
+    @ResponseBody
     @RequestMapping(value = Lists.ClassList + "/list", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
-    public @ResponseBody Object getAgencyList(HttpServletRequest request) throws Exception {
+    public Object getAgencyList(HttpServletRequest request) throws Exception {
         long total = comService.getCount(BookClassification.class);
         int tarPageNum = Integer.parseInt(request.getParameter("tarPageNum"));
         int perPageNum = Integer.parseInt(request.getParameter("perPageNum"));
         List<BookClassification> bcs = comService.getList(BookClassification.class, tarPageNum, perPageNum);
         List<ClassList> list = new ArrayList<>();
-        for (BookClassification bc: bcs) {
+        for (BookClassification bc : bcs) {
             list.add(new ClassList(bc));
         }
         Map<String, Object> result = new HashMap<>();
@@ -54,15 +56,17 @@ public class BookClassController {
         return Result.SUCCESS(result);
     }
 
+    @ResponseBody
     @RequestMapping(value = Forms.ClassForm + "/form", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
-    public @ResponseBody Object getAgencyForm(HttpServletRequest request) throws Exception {
+    public Object getAgencyForm(HttpServletRequest request) throws Exception {
         String dataId = request.getParameter("dataId");
         BookClassification bc = comService.getDetail(BookClassification.class, Integer.parseInt(dataId));
         return Result.SUCCESS(new ClassList(bc));
     }
 
+    @ResponseBody
     @RequestMapping(value = Forms.ClassForm + "/submit", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
-    public @ResponseBody Object submitAgency(HttpServletRequest request) throws Exception {
+    public Object submitAgency(HttpServletRequest request) throws Exception {
         String data = request.getParameter("data");
         ClassList form = DataManager.string2Object(data, ClassList.class);
         if (form == null) {
@@ -76,29 +80,34 @@ public class BookClassController {
             BookClassification bc = new BookClassification(form);
             comService.saveDetail(bc);
             cacheManager.resetBookClassificationCache();
-            comService.saveDetail(new AdminLog(adminSession, OperationTargets.BookClassification, OperationTypes.Create, String.valueOf(bc.getId()), "分类名称： "+bc.getName()));
+            comService.saveDetail(new AdminLog(adminSession, OperationTargets.BookClassification, OperationTypes.Create, String.valueOf(bc.getId()), "分类名称： " + bc.getName()));
             return Result.SUCCESS(bc.getId());
         } else {
-            if (comService.hasExist(BookClassification.class, "name='" + form.getName() + "' and id!="+Integer.parseInt(form.getId()))) {
+            if (comService.hasExist(BookClassification.class, "name='" + form.getName() + "' and id!=" + Integer.parseInt(form.getId()))) {
                 return Result.ERROR(ErrorCode.CUSTOMIZED_ERROR, "分类名已存在");
             }
             BookClassification bc = new BookClassification(form);
             bc.setId(Integer.parseInt(form.getId()));
             comService.saveDetail(bc);
             cacheManager.resetBookClassificationCache();
-            comService.saveDetail(new AdminLog(adminSession, OperationTargets.BookClassification, OperationTypes.Update, String.valueOf(bc.getId()), "分类名称： "+bc.getName()));
+            comService.saveDetail(new AdminLog(adminSession, OperationTargets.BookClassification, OperationTypes.Update, String.valueOf(bc.getId()), "分类名称： " + bc.getName()));
             return Result.SUCCESS();
         }
     }
 
+    @ResponseBody
     @RequestMapping(value = Lists.ClassList + "/delete", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
-    public @ResponseBody Object deleteAgency(HttpServletRequest request) throws Exception {
+    public Object deleteAgency(HttpServletRequest request) throws Exception {
         String dataId = request.getParameter("dataId");
+        Book book = comService.getFirst(Book.class, "classificationId="+dataId);
+        if (book!=null){
+            return Result.ERROR(ErrorCode.CUSTOMIZED_ERROR, "该分类下还有相关图书，无法删除");
+        }
         BookClassification bc = comService.getDetail(BookClassification.class, Integer.parseInt(dataId));
         comService.deleteDetail(bc);
         cacheManager.resetBookClassificationCache();
         AdminSession adminSession = AdminValidator.getAdminSession(request);
-        comService.saveDetail(new AdminLog(adminSession, OperationTargets.BookClassification, OperationTypes.Delete, String.valueOf(bc.getId()), "分类名称： "+bc.getName()));
+        comService.saveDetail(new AdminLog(adminSession, OperationTargets.BookClassification, OperationTypes.Delete, String.valueOf(bc.getId()), "分类名称： " + bc.getName()));
         return Result.SUCCESS();
     }
